please pass more tests
git-svn-id: file:///svn/unbound/branches/edns-subnet@2967 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ edns-subnet/addrtree.c
@@ -29,7 +29,7 @@ edge_create(struct addrnode* node, const addrkey_t* addr, addrlen_t addrlen)
 		return NULL;
 	edge->node = node;
 	edge->len = addrlen;
-	n = (addrlen / KEYWIDTH) + ((addrlen % KEYWIDTH)!=0)?1:0; /*ceil()*/
+	n = (size_t)((addrlen / KEYWIDTH) + ((addrlen % KEYWIDTH)!=0)?1:0); /*ceil()*/
 	edge->str = (addrkey_t*)calloc(n, sizeof(addrkey_t));
 	if (!edge->str) {
 		free(edge);
